Added an --ignore option
This filters packages with no valid versions out of the generated date.
This makes it easier to debug things (since there isn't a lot of junk
in the output file).  But I think it would be good to include such packages
by default simply because 'search' will pick them up which means that people
will become aware of packages and then potentially push the developers to
use the proper conventions to allow those packages to be install via
impact.