Finally able to close issue #1
This implements a very simple scheme for loading dependencies.  It doesn't
try to do the more sophisticated thing of finding a set of versions of
libraries that are all mutually compatible.  Instead, it starts from a
base library and then determines the specific dependencies of that
library.

If Modelica ever adopts a version annotation that supports compatibility
with multiple libraries, this version elaboration process will get more
complicated (but also more powerful).  But for now, we at least have this
simple approach.

This closes #1.