Push the pagination limit of GitHub to max(=100)
For some strange reason also the API results in GitHub are
paginated. This means by default we only get the first results.
This patch sets the number of results per "page" up to 100
which is the current maximum. If this is not enough we have to
start making several requests pulling the different "pages".