Better handling of pre-releases and build metadata
This implements the approach that @dietmarw suggested in ticket #29.  Specifically,
it refines the way that we look for the library in the repository by stripping
any pre-release or build information off, looking for a matching directory but then
storing the version information under the full version id (i.e. with the pre-release
and/or build data included).

I tested this and it seems to work so this closes #29.