Applying some best practices
A good practice with Go (or any other language really) is to avoid defining multiple
types in a single file.  These changes break up previously monolithic files into
separate files.  This makes it easier to find things and it helps to document the
code better since the directories and files represent the packages and types.