Merged back in the list command and recommented one assertion
The next step is going to be to remove references to dependency resolution in the index
package and, instead, depend (no pun intended) on the graph package to resolve dependencies.
The commandline tools (install) already do this.  But it needs to be in the library side,
not just the command line part.  I'm also going to switch everything over to using semantic
versions globally.