Start of a crawler package
This time around, I want to build more flexibility into this system.  As such, I'm
creating abstractions for both the things that crawl repositories and things that
record the results.  My hope is that we can then implement crawlers that can
crawl not just GitHub repos, but also private repositories (e.g., svn).  I've
created the Recorder interface so that we don't necessarily build JSON from the
start but perhaps put it into a database of some kind which we then generate JSON
from.  The point here is that if we want to support something like WebHooks
to update just parts of the index, it would be better to have a database behind
the scenes.  But for a local user wanting to generate a local index, we might
simply go straight to JSON.  So I just want to build in some flexibility.