More work on indexing
I'm still working mainly on the crawling part.  The GitHub API is rich with information
but we need to translate that into something more generic.

I added a new package called 'dirinfo'.  The idea there is that someone creating a library
can include an 'impact.json' file that provides information about what libraries are
present.  The idea is that we will try to infer this information (i.e., if such a file
is not present, we will automatically build an inferred version).  But for complex
cases, developers can use the 'impact.json' file to explicitly specify information.