Running gh_crawl_test generates an index, but much more work to do
I'm not picking up nearly enough information.  But this is a good start.

One thing I still need to think about is how to bring non-semantic versioned libraries
into this.  One way would be to use the impact.json file to somehow transform things.
Another would be to assume some convention for mapping back and forth.