Doing a decent job of crawling
I'm starting to add support for reading impact.json from repositories to help
in some situations.  I've also added some additional tests and some additional
heuristics for repositories where the Modelica package is at the root of the
repository and cases where packages are stored as directories with version
numbers AND the tag includes extra 'junk'.