All information is getting filled in
There is one issue remaining which I tried to address here but isn't working.  If
a repository is a fork, we should really use information from the original library,
not the fork.  I tried to do that by following the Source repository, but it didn't
seem to work.  I need to investigate why.  I think this is why email isn't working
since the owner is an organization that doesn't appear to have an email.