This version pulls a complete representation of the repository, but doesn't actually use it yet
The point here is that we want access to the 'source' and 'parent' elements of the repository
which you do not get with a simple 'List' call.  So I then 'Get' the repository as well
so I get complete information.  But I have not yet decided what I intend to do with that
information (see #94).