A big push to work on installation related functionality
This includes work to load and manipulate the index as well as the code to build the
dependency graph and resolve dependencies.  In fact, the 'install' command actually
seems to work at this point.  It uses the sample index in the root impact package for
dependencies, build the dependency graph and resolves the dependencies.  It just doesn't
install anything yet.