I think this can handle forks and duplicates in a reasonable way
I think this should handle forks and duplicates, but it is untested.  I need to make some
changes to the indexing so that we actually have some forks (e.g., Buildings) and then
run through the functionality of 'install' and 'search' to see how it works.  Once that is
done, it should resolve #80.