This adds extensive discussion of file formats
I'm a bit torn on two points.  First, should impact.json really be JSON or Denada.  Also,
I didn't document impact.proj so any documentation about that seems like it would be out
of order.  Perhaps the schema information should be moved to the end as an appendix and
should include the impact.proj stuff (in Denada format?)