Wrote up quite a bit about dependency resolution
I liked this approach initially because it introduced things somewhat gradually.  After
finishing it, I wonder if it isn't too 'meandering'.  I really need to have a definitive
way of stating the algorithm...not sure how to do that.

The last example demonstrates what I wanted to show, but I wonder if it could be better.
I created a graphic to go with it that *definitely* can be better.  But I wonder how to
reconcile all these different ways of discussing the algorithm.  Perhaps they all have
merit for explaining different aspects of the algorithm or just because some may be
easier to follow than others.  In the end though, I really need a way to succinctly
describe the algorithm in case others wish to implement it.  Hmmm....